Automatically lock after inactivity
This can be configured with 5 clicks from lockout mode. Click the number
of minutes after the flashlight should automatically switch into lockout
mode. If you don't click at all it will turn off this feature (this is
in contrast to other config modes that keep their current value if the
user did not click, but it saves some flash).