Fix output configuration
Although you can choose the pin for each output in the hw config file,
fsm expects a fixed channel/pin mapping. This commit changes the output
configuration from hardcoded to hardcoded and is only compatible with
the hardware configuration of the Astrolux FT03.

TODO: Automatically configure the outputs as necessary.