added USE_2C_MAX_TURBO option from Tom E
Makes "Ramp 2C" behave like Anduril1 instead of Anduril2.
So it goes to full power turbo in the advanced UI, or ceiling in the simple UI.
Otherwise, default Anduril2 behavior is for "Ramp 2C" to go to/from the ceiling
level... unless already ramped up to that level in advanced mode, in which case
it'll go to full turbo.