added #define USE_THERM_AUTOCALIBRATE to opt out
Moved factory reset thermal auto-calibration behind a new default-on
config #define USE_THERM_AUTOCALIBRATE.  Commenting this out or the
usual #ifdef/#undef in a cfg-[...].h build file allows for manually
calibrating the temperature offset.

This may be useful for factory-calibrated temperature sensors or for
those who regularly flash custom builds and don't want to recalibrate
each time.

Determining the correct temperature offset for a given flashlight
first requires flashing a build with auto-calibrate disabled, using
that to determine the offset, which can then be baked in to future
firmware builds.