set KR4 default floor higher at Hank's request, and reduced the jump-start-moon power a bit too (he says both are good now, but I don't have hardware to measure it myself)
The higher floor is because, when the driver is really hot from being on turbo,
going directly to moon causes the LEDs to turn off for a while until the driver cools.
The new floor is the lowest level where post-turbo activation works reliably.
However, it should turn on even at the level 1/150 when it's not hot.