Add option to stop ramping up from moon
With `#undef MOON_START_RAMPING` entering moon from off by holding the
switch will not automatically start ramping up. This ensures that moon
is entered reliably when the switch is held long enough.

Timings still apply (you can release the switch when the light turns on
or when it blips, depending on the configured timing).