Partially fixed oscillating aux LED voltage colors while asleep.
Amplitude is smaller, frequency is slower, but it can still happen sometimes at
color boundaries on models with bright aux LEDs on high mode.  (because that's
not a measurement error, it's actually oscillating voltage)

The Wurkkos TS11 in particular tends to bounce when crossing certain color
boundaries, because the aux LEDs are bright and change the voltage enough
to push it back and forth across the boundary.

Also sped up voltage measurement during first few seconds after turning off,
to compensate for slowdown caused by the lowpass filter... while slowing down
measurements afterward to slow the oscillation.

I tried a bunch of different methods, spanning a range of "stable but slow" to
"fast but unstable", and kept code clauses for three representative methods.
The one enabled by default is the best compromise I found to reduce the issue
as much as possible while keeping readings reasonably responsive.